goruntuisleme/legacy_python/requirements.txt
opencv-python
ultralytics
torch
torchvision
albumentations
numpy
tqdm
icrawler
python-dotenv
openai
